解决feign调用 insert方法的没有返回最新数据id问题

+++ ucs-service-api/ucs-user-api/src/main/java/com/sailmi/system/user/feign/IUserClient.java
@@ -63,5 +63,5 @@ public interface IUserClient {
 	 * @return
 	 */
 	@PostMapping(API_PREFIX + "/submit-user")
-	public int  submitUserInfo(@RequestBody User user);
+	public Long  submitUserInfo(@RequestBody User user);
 }
